+++ core/src/com/t2wonderland/kurona/interfaces/IGameObject.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.math.Rectangle;
 import com.badlogic.gdx.math.Vector2;
 
 public interface IGameObject {
-    void setInitialPosition(int x, int y);
+    void setInitialPosition(float x, float y);
     Vector2 getSize();
     Vector2 getPosition();
     Rectangle getBounds();
